added parallel concept to client and fixed a typo in code and fixed an error which caused timers to be less precise

+++ Libraries/graphics/graphics.cpp
@@ -173,7 +173,7 @@ void gfx::blurTexture(sf::RenderTexture& texture, float blur_intensity) {
 }
 
 float gfx::Timer::getTimeElapsed() const {
-    return clock.getElapsedTime().asMicroseconds() / 1000;
+    return (float)clock.getElapsedTime().asMicroseconds() / 1000;
 }
 
 void gfx::Timer::reset() {
